+++ src/AppWrapper.js
@@ -3,6 +3,7 @@ import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
 import dhis2theme from '@dhis2/d2-ui-core/theme/mui3.theme'
 import { Provider as ReduxProvider } from 'react-redux'
 import { D2Shim } from '@dhis2/app-runtime-adapter-d2'
+import { useDataEngine } from '@dhis2/app-runtime'
 
 import WindowDimensionsProvider from './components/WindowDimensionsProvider'
 import App from './components/App'
@@ -35,8 +36,10 @@ if (authorization) {
 }
 
 const AppWrapper = () => {
+    const dataEngine = useDataEngine()
+
     return (
-        <ReduxProvider store={configureStore()}>
+        <ReduxProvider store={configureStore(dataEngine)}>
             <MuiThemeProvider theme={muiTheme()}>
                 <D2Shim d2Config={d2Config} i18nRoot="./i18n">
                     {({ d2 }) => {
@@ -46,7 +49,7 @@ const AppWrapper = () => {
                         }
                         return (
                             <WindowDimensionsProvider>
-                                <App d2={d2} />
+                                <App d2={d2} dataEngine={dataEngine} />
                             </WindowDimensionsProvider>
                         )
                     }}