+++ src/components/Item/VisualizationItem/__tests__/ItemHeaderButtons.spec.js
@@ -8,7 +8,7 @@ jest.mock('../Visualization/plugin', () => ({
     pluginIsAvailable: () => true,
 }))
 
-it('renders correctly', () => {
+it('renders correctly when not fullscreen', () => {
     const buttons = shallow(
         <ItemHeaderButtons
             item={{
@@ -23,6 +23,27 @@ it('renders correctly', () => {
             activeType={'CHART'}
             d2={{}}
             onToggleFooter={Function.prototype}
+        />
+    )
+    expect(toJson(buttons)).toMatchSnapshot()
+})
+
+it('renders correctly when fullscreen', () => {
+    const buttons = shallow(
+        <ItemHeaderButtons
+            item={{
+                type: 'CHART',
+                chart: { type: 'NOT_YOY', domainType: 'AGGREGATE' },
+            }}
+            visualization={{
+                type: 'SINGLE_VALUE',
+            }}
+            onSelectActiveType={Function.prototype}
+            activeFooter={false}
+            activeType={'CHART'}
+            d2={{}}
+            onToggleFooter={Function.prototype}
+            isFullscreen={true}
         />
     )
     expect(toJson(buttons)).toMatchSnapshot()